Renamed sqs-source-new to sqs-source and removed deprecated sqs-source implementation
