--- data-prepper-plugins/sqs-source/src/integrationTest/java/org/opensearch/dataprepper/plugins/source/sqssource/RecordsGenerator.java
@@ -1,11 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-package org.opensearch.dataprepper.plugins.source.sqssource;
-
-import java.util.List;
-
-public interface RecordsGenerator {
-    void pushMessages(final List<String> messages, final String queueUrl);
-}
