+++ data-prepper-plugins/sqs-source/src/main/java/org/opensearch/dataprepper/plugins/source/sqs/SqsService.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
- package org.opensearch.dataprepper.plugins.source.sqssourcenew;
+ package org.opensearch.dataprepper.plugins.source.sqs;
 
  import com.linecorp.armeria.client.retry.Backoff;
  import org.opensearch.dataprepper.common.concurrent.BackgroundThreadFactory;
